requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
SQLAlchemy==2.0.21
alembic==1.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyserial==3.5
redis==5.0.1
rq==1.15.1
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
